+++ src/thread_factory/primitives/__init__.py
@@ -2,10 +2,12 @@ from thread_factory.primitives.dynaphore import Dynaphore
 from thread_factory.primitives.smart_condition import SmartCondition
 from thread_factory.primitives.switchlock import SwitchLock
 from thread_factory.primitives.signal_condition import SignalCondition
+from thread_factory.primitives.signal_barrier import SignalBarrier
 
 __all__ = [
     "Dynaphore",
     "SmartCondition",
     "SwitchLock",
     "SignalCondition",
+    "SignalBarrier",
     ]