+++ src/thread_factory/utils/coordination/__init__.py
@@ -1,5 +1,7 @@
 from thread_factory.utils.coordination.group import Group
+from thread_factory.utils.coordination.router_group import RouterGroup
 
 __all__ = [
     "Group",
+    "RouterGroup",
 ]